+++ src/app/chatbot-design-studio/utils.ts
@@ -112,6 +112,7 @@ export enum TYPE_INTENT_NAME {
     CLOSE              = "close"
 }
 
+
 export enum TYPE_MATH_OPERATOR {
     addAsNumber         = "addAsNumber",
     addAsString         = "addAsString",